Sentiment Analysis App Streamlit/test/huggingface.py
from langchain_huggingface import HuggingFaceEndpoint,ChatHuggingFace

llm = HuggingFaceEndpoint(
    repo_id="microsoft/Phi-3-mini-4k-instruct",
    task="text-generation",
    max_new_tokens=512,
    do_sample=False,
    repetition_penalty=1.03,
)

chat = ChatHuggingFace(llm=llm, verbose=True)
messages = [("system", "You are a helpful translator. Translate the user sentence to French."),
    ("human", "I love programming."),
]

chat.invoke(messages)