Sentiment Analysis App Streamlit/requirements.txt
streamlit==1.40.2
pandas
langchain-core==0.3.71
langchain-groq==0.3.6
langchain==0.3.26
python-dotenv==1.0.0
blinker==1.9.0
protobuf
google
typing_extensions
cachetools
click
tornado
toml
numpy
pytz
python-dateutil
six
pydantic_core
typing_inspection
pydantic
pydantic_core